requirements.txt
neptune-client
neptune-sklearn
numpy
pandas
scikit-learn
scikit-optimize
wandb